+++ packages/account-tree-controller/src/index.ts
@@ -1,6 +1,7 @@
 export type { AccountWalletObject } from './wallet';
+export { isMultichainAccountWallet } from './wallet';
 export type { AccountGroupObject } from './group';
-export { isAccountGroupNameUnique } from './group';
+export { isAccountGroupNameUnique, isMultichainAccountGroup } from './group';
 
 export {
   USER_STORAGE_GROUPS_FEATURE_KEY,