+++ packages/account-tree-controller/tests/mockMessenger.ts
@@ -49,6 +49,7 @@ export function getAccountTreeControllerMessenger(
       'AccountsController:selectedAccountChange',
       'UserStorageController:stateChange',
       'MultichainAccountService:walletStatusChange',
+      'MultichainAccountService:groupStatusChange',
     ],
     actions: [
       'AccountsController:listMultichainAccounts',
@@ -61,6 +62,8 @@ export function getAccountTreeControllerMessenger(
       'UserStorageController:performSetStorage',
       'UserStorageController:performBatchSetStorage',
       'AuthenticationController:getSessionProfile',
+      'MultichainAccountService:getMultichainAccountGroup',
+      'MultichainAccountService:getMultichainAccountWallet',
       'MultichainAccountService:createMultichainAccountGroup',
       'KeyringController:getState',
       'SnapController:getSnap',